refactor: consolidate duplicated model config, usage extraction, and daemon helpers
- Extract ModelConfigFields with merge_from()/or_fallback(), eliminating 3x18
  duplicated field declarations across ProviderConfig/ModelEntry/ResolvedModel
- Add ResolvedModel::from_parts() constructor, replacing 18 inline .or() calls
- Replace 3 identical validate_model_ref() calls with loop; add missing
  defaults.collation validation
- Fix CompactionConfig::idle_trigger_minutes type mismatch (u64→u32), remove
  3 unsafe casts
- Fix compact() model resolution: use resolve_agent_model() full fallback chain
  instead of incomplete active_model-only check
- Add open_embed_and_vectorstore() helper, consolidating 2 inline vector store
  setup blocks
- Extract merge_anthropic_usage(), extract_anthropic_usage(),
  extract_gemini_usage() into stream_helpers, replacing 5 inline extraction sites

+++ shore-daemon/src/main.rs
@@ -363,7 +363,7 @@ async fn run_compaction(
 
     let app_compaction = &effective.app.memory.compaction;
     let mgr_config = CompactionConfig {
-        idle_trigger_minutes: app_compaction.idle_trigger_minutes as u64,
+        idle_trigger_minutes: app_compaction.idle_trigger_minutes,
         min_turns: app_compaction.min_turns,
         max_turns: app_compaction.max_turns,
         keep_recent_turns: app_compaction.keep_recent_turns,
